Levy Transaction class test and method migrations.
