+++ src/SFA.DAS.EAS.Web/Orchestrators/EmployerAccountPayeOrchestrator.cs
@@ -16,7 +16,6 @@ using SFA.DAS.EAS.Application.Queries.GetTeamUser;
 using SFA.DAS.EAS.Domain.Configuration;
 using SFA.DAS.EAS.Domain.Interfaces;
 using SFA.DAS.EAS.Domain.Models.Account;
-using SFA.DAS.EAS.Domain.Models.UserProfile;
 using SFA.DAS.EAS.Web.ViewModels;
 using SFA.DAS.EAS.Web.ViewModels.AccountPaye;
 using SFA.DAS.NLog.Logger;
