+++ src/SFA.DAS.EAS.Web/Orchestrators/EmployerAccountTransactionsOrchestrator.cs
@@ -1,14 +1,9 @@
 ﻿using MediatR;
 using SFA.DAS.EAS.Application.Exceptions;
-using SFA.DAS.EAS.Application.Queries.FindAccountCoursePayments;
 using SFA.DAS.EAS.Application.Queries.FindAccountProviderPayments;
-using SFA.DAS.EAS.Application.Queries.FindEmployerAccountLevyDeclarationTransactions;
 using SFA.DAS.EAS.Application.Queries.GetEmployerAccount;
 using SFA.DAS.EAS.Application.Queries.GetEmployerAccountTransactions;
-using SFA.DAS.EAS.Application.Queries.GetPayeSchemeByRef;
-using SFA.DAS.EAS.Domain.Extensions;
 using SFA.DAS.EAS.Domain.Interfaces;
-using SFA.DAS.EAS.Domain.Models.Levy;
 using SFA.DAS.EAS.Domain.Models.Transaction;
 using SFA.DAS.EAS.Web.Models;
 using SFA.DAS.EAS.Web.ViewModels;
@@ -18,7 +13,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using SFA.DAS.Activities.Client;
 using TransactionLine = SFA.DAS.EAS.Domain.Models.Transaction.TransactionLine;
 
 namespace SFA.DAS.EAS.Web.Orchestrators
@@ -40,43 +34,7 @@ namespace SFA.DAS.EAS.Web.Orchestrators
 			_currentTime = currentTime;
 			_logger = logger;
 		}
-        /// <summary>
-        /// AML-2454: Move to Finance orchestrator as local Controller no longer processes this feature
-        /// </summary>
-		public async Task<OrchestratorResponse<TransactionLineViewModel<LevyDeclarationTransactionLine>>>
-			FindAccountLevyDeclarationTransactions(
-				string hashedId, DateTime fromDate, DateTime toDate, string externalUserId)
-		{
-			var data = await _mediator.SendAsync(new FindEmployerAccountLevyDeclarationTransactionsQuery
-			{
-				HashedAccountId = hashedId,
-				FromDate = fromDate,
-				ToDate = toDate,
-				ExternalUserId = externalUserId
-			});
-
-			foreach (var transaction in data.Transactions)
-			{
-				var payeSchemeData = await _mediator.SendAsync(new GetPayeSchemeByRefQuery
-				{
-					HashedAccountId = hashedId,
-					Ref = transaction.EmpRef
-				});
-
-				transaction.PayeSchemeName = payeSchemeData?.PayeScheme?.Name ?? string.Empty;
-			}
-
-			return new OrchestratorResponse<TransactionLineViewModel<LevyDeclarationTransactionLine>>
-			{
-				Status = HttpStatusCode.OK,
-				Data = new TransactionLineViewModel<LevyDeclarationTransactionLine>
-				{
-					Amount = data.Total,
-					SubTransactions = data.Transactions,
-					TransactionDate = data.Transactions.First().DateCreated
-				}
-			};
-		}
+       
 
 		public async Task<OrchestratorResponse<PaymentTransactionViewModel>> FindAccountPaymentTransactions(
 			string hashedId, long ukprn, DateTime fromDate, DateTime toDate, string externalUserId)
