+++ src/SFA.DAS.EmployerApprenticeshipsService.Application.UnitTests/SFA.DAS.EAS.Application.UnitTests.csproj
@@ -389,8 +389,6 @@
     <Compile Include="Queries\GetAccountTasksTests\WhenIGetAnAccountsTasks.cs" />
     <Compile Include="Queries\GetAccountTasksTests\WhenIValidateTheRequest.cs" />
     <Compile Include="Queries\GetAccountTeamMembers\WhenIValidateTheRequest.cs" />
-    <Compile Include="Queries\GetAccountTransactionDetailTests\WhenIGetAccountTransactionDetails.cs" />
-    <Compile Include="Queries\GetAccountTransactionDetailTests\WhenIValidateTheRequest.cs" />
     <Compile Include="Queries\GetAccountTransactionSummaryTests\WhenIGetAccountTransactionSummary.cs" />
     <Compile Include="Queries\GetAccountTransactionsTests\WhenIGetAccountTransactions.cs" />
     <Compile Include="Queries\GetAccountTransactionsTests\WhenIValidateTheRequest.cs" />
@@ -413,10 +411,6 @@
     <Compile Include="Queries\GetEmployerAccountTests\WhenIGetEmployerAccountsByHashedKey.cs" />
     <Compile Include="Queries\GetEmployerAccountTests\WhenIValidateTheGetAccountByHashedIdRequest.cs" />
     <Compile Include="Queries\GetEmployerAccountTests\WhenIValidateTheGetAccountByIdRequest.cs" />
-    <Compile Include="Queries\FindEmployerAccountPaymentTransactionDetailsTests - Copy\WhenIGetEmployerPaymentTransactionDetails.cs" />
-    <Compile Include="Queries\FindEmployerAccountPaymentTransactionDetailsTests - Copy\WhenIValidateTheRequest.cs" />
-    <Compile Include="Queries\FindEmployerAccountLevyDeclarationTransactionDetailsTests\WhenIGetEmployerLevyDeclarationTransactionDetails.cs" />
-    <Compile Include="Queries\FindEmployerAccountLevyDeclarationTransactionDetailsTests\WhenIValidateTheRequest.cs" />
     <Compile Include="Queries\GetEmployerAccountTransactionsTests\WhenIGetEmployerTransactions.cs" />
     <Compile Include="Queries\GetEmployerAccountTransactionsTests\WhenIValidateTheRequest.cs" />
     <Compile Include="Queries\GetEmployerAgreementTemplateTests\WhenIRequestTheLatestEmployerAgreementTemplate.cs" />
@@ -469,7 +463,6 @@
     <Compile Include="Services\DasLevyServiceTests\WhenIGetEnglishFractionForAnEmpref.cs" />
     <Compile Include="Services\DasLevyServiceTests\WhenIGetPreviousTransactionCount.cs" />
     <Compile Include="Services\DasLevyServiceTests\WhenIGetTheLastDeclarationForAPayeRef.cs" />
-    <Compile Include="Services\DasLevyServiceTests\WhenIGetTransactionDetailInformation.cs" />
     <Compile Include="Services\DasLevyServiceTests\WhenIGetTransactionInformation.cs" />
     <Compile Include="Services\RefreshEmployerLevyServiceTests\WhenQueuingAMessage.cs" />
   </ItemGroup>
