+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/GetPayeSchemeByRef/GetPayeSchemeByRefResponse.cs
@@ -2,9 +2,6 @@
 
 namespace SFA.DAS.EAS.Application.Queries.GetPayeSchemeByRef
 {
-    /// <summary>
-    ///  AML-2454: Copy to finance
-    /// </summary>
     public class GetPayeSchemeByRefResponse
     {
         public PayeSchemeView PayeScheme { get; set; }
