+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Data/Repositories/ITransactionRepository.cs
@@ -9,9 +9,7 @@ namespace SFA.DAS.EAS.Domain.Data.Repositories
     public interface ITransactionRepository
     {
         Task<List<TransactionLine>> GetAccountTransactionsByDateRange(long accountId, DateTime fromDate, DateTime toDate);
-
-        Task<List<TransactionLine>> GetAccountLevyTransactionsByDateRange(long accountId, DateTime fromDate, DateTime toDate);
-
+        
         Task<List<TransactionLine>> GetAccountTransactionByProviderAndDateRange(long accountId, long ukprn, DateTime fromDate, DateTime toDate);
 
         Task<List<TransactionLine>> GetAccountCoursePaymentsByDateRange(long accountId, long ukprn, string courseName, int? courseLevel, int? pathwayCode, DateTime fromDate, DateTime toDate);
