+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Interfaces/IDasLevyService.cs
@@ -11,13 +11,6 @@ namespace SFA.DAS.EAS.Domain.Interfaces
     {
         Task<ICollection<TransactionLine>> GetAccountTransactionsByDateRange(long accountId, DateTime fromDate, DateTime toDate);
 
-        /// <summary>
-        ///  AML-2454: Move to finance
-        /// </summary>
-        Task<ICollection<T>> GetAccountLevyTransactionsByDateRange<T>(
-            long accountId, DateTime fromDate, DateTime toDate)
-            where T : TransactionLine;
-
         Task<ICollection<T>> GetAccountProviderPaymentsByDateRange<T>(
             long accountId, long ukprn, DateTime fromDate, DateTime toDate)
             where T : TransactionLine;
