+++ src/SFA.DAS.EmployerFinance/Data/ITransactionRepository.cs
@@ -14,5 +14,7 @@ namespace SFA.DAS.EmployerFinance.Data
         Task<List<TransactionLine>> GetAccountTransactionsByDateRange(
             long accountId, DateTime fromDate, DateTime toDate);
  		Task<List<TransactionLine>> GetAccountCoursePaymentsByDateRange(long accountId, long ukprn, string courseName, int? courseLevel, int? pathwayCode, DateTime fromDate, DateTime toDate);
+        Task<List<TransactionLine>> GetAccountLevyTransactionsByDateRange(long accountId, DateTime fromDate, DateTime toDate);
+
     }
 }
