+++ src/SFA.DAS.EmployerFinance/Queries/FindEmployerAccountLevyDeclarationTransactions/FindEmployerAccountLevyDeclarationTransactionsResponse.cs
@@ -1,11 +1,8 @@
 ﻿using System.Collections.Generic;
-using SFA.DAS.EAS.Domain.Models.Levy;
+using SFA.DAS.EmployerFinance.Models.Levy;
 
-namespace SFA.DAS.EAS.Application.Queries.FindEmployerAccountLevyDeclarationTransactions
+namespace SFA.DAS.EmployerFinance.Queries.FindEmployerAccountLevyDeclarationTransactions
 {
-    /// <summary>
-    ///  AML-2454: Move to finance
-    /// </summary>
     public class FindEmployerAccountLevyDeclarationTransactionsResponse
     {
         public List<LevyDeclarationTransactionLine> Transactions { get; set; }
