+++ src/SFA.DAS.EmployerFinance/Queries/GetAccountLevyTransactions/GetAccountLevyTransactionsQueryValidator.cs
@@ -2,11 +2,8 @@
 using System.Threading.Tasks;
 using SFA.DAS.Validation;
 
-namespace SFA.DAS.EAS.Application.Queries.AccountTransactions.GetAccountLevyTransactions
+namespace SFA.DAS.EmployerFinance.Queries.GetAccountLevyTransactions
 {
-    /// <summary>
-    ///  AML-2454: Move to finance
-    /// </summary>
     public class GetAccountLevyTransactionsQueryValidator : IValidator<GetAccountLevyTransactionsQuery>
     {
         public ValidationResult Validate(GetAccountLevyTransactionsQuery item)
