+++ src/SFA.DAS.EmployerFinance/Queries/GetAccountLevyTransactions/GetAccountLevyTransactionsResponse.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using SFA.DAS.EmployerFinance.Models.Transaction;
+
+namespace SFA.DAS.EmployerFinance.Queries.GetAccountLevyTransactions
+{
+    public class GetAccountLevyTransactionsResponse
+    {
+        public List<TransactionLine> Transactions { get; set; }
+    }
+}
+
