+++ src/SFA.DAS.EmployerFinance/Queries/GetPayeSchemeByRef/GetPayeSchemeByRefResponse.cs
@@ -1,0 +1,9 @@
+﻿using SFA.DAS.EmployerFinance.Models.Account;
+
+namespace SFA.DAS.EmployerFinance.Queries.GetPayeSchemeByRef
+{
+   public class GetPayeSchemeByRefResponse
+    {
+        public PayeSchemeView PayeScheme { get; set; }
+    }
+}
